do a bit of restructuring for better consistency
