+++ backend/user-service/src/__tests__/unit/delete-user.test.ts
@@ -1,5 +1,5 @@
 import supertest from "supertest";
-import HttpStatusCode from "../../lib/HttpStatusCode";
+import HttpStatusCode from "../../lib/enums/HttpStatusCode";
 import db from "../../lib/db";
 import createServer from "../utils/server";
